new analysis: cache latency
Adding new code that test the cache latency with the pointer chasing
function.

It work by creating an array of pointer pointing on themself,
shuffle it and iterate through the array.

When the array wan fit in the L1 cache, the time to acces the data is
theoricaly the time to acces the L1 cache.

same raisoning for the L2 and L3 cache